+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 ## v0.0.14 - In progress
 ### Added
 - Redirect to the correct entity type, if we've landed on the wrong one.
+- Add provenance table to details.
+- Add expansion panels for details data.
+- Add prop types checks to components.
+### Changed
+- Make details page react by default.
+- Remove unused templates.
+- Fix data in details top title panel.
+- Remove breadcrumbs links from details top title panel.
 
 ## [v0.0.13](https://github.com/hubmapconsortium/portal-ui/tree/v0.0.13) - 2020/05/03
 ### Added